+++ packages/nimble-components/src/wafer-map/modules/experimental/hover-handler.ts
@@ -1,5 +1,5 @@
-import type { WaferMap } from '..';
-import { PointCoordinates, WaferMapOriginLocation } from '../types';
+import type { WaferMap } from '../..';
+import { PointCoordinates, WaferMapOriginLocation } from '../../types';
 import { DataManager } from './data-manager';
 
 /**
@@ -24,26 +24,20 @@ export class HoverHandler {
         this.wafermap.removeEventListener('mouseout', this.onMouseOut);
     }
 
-    private readonly onMouseMove = (event: MouseEvent): void => {
-        if (this.wafermap.diesTable !== undefined) {
-            return;
-        }
-        const mousePosition: PointCoordinates = {
-            x: event.offsetX,
-            y: event.offsetY
-        };
-
-        if (!this.hoversOverDie(this.wafermap, mousePosition)) {
-            this.wafermap.hoverDie = undefined;
+    /**
+     * @internal
+     * keep public for testing until data manager refactor
+     */
+    public readonly onMouseMove = (event: MouseEvent): void => {
+        if (this.wafermap.diesTable === undefined) {
             return;
         }
         // get original mouse position in case we are in zoom.
         const invertedPoint = this.wafermap.transform.invert([
-            mousePosition.x,
-            mousePosition.y
+            event.offsetX,
+            event.offsetY
         ]);
-
-        const dieCoordinates = this.calculateDieCoordinates(this.wafermap, {
+        const dieCoordinates = this.calculateDieCoordinates({
             x: invertedPoint[0],
             y: invertedPoint[1]
         });
@@ -51,10 +45,28 @@ export class HoverHandler {
             this.wafermap.hoverDie = undefined;
             return;
         }
+        const colIndex = this.wafermap.diesTable
+            .getChild('colIndex')!
+            .toArray() as Int32Array;
+        const rowIndex = this.wafermap.diesTable
+            .getChild('rowIndex')!
+            .toArray() as Int32Array;
 
-        if (this.wafermap.dataManager instanceof DataManager) {
-            this.wafermap.hoverDie = this.wafermap.dataManager.getWaferMapDie(dieCoordinates);
+        // will replace iterating with arquero filtering after fixing errors
+        for (let i = 0; i < colIndex.length; i++) {
+            if (
+                colIndex[i] === dieCoordinates.x
+                && rowIndex[i] === dieCoordinates.y
+            ) {
+                this.wafermap.hoverDie = {
+                    index: i,
+                    x: dieCoordinates.x,
+                    y: dieCoordinates.y
+                };
+                return;
+            }
         }
+        this.wafermap.hoverDie = undefined;
     };
 
     private readonly onMouseOut = (_event: MouseEvent): void => {
@@ -62,49 +74,31 @@ export class HoverHandler {
     };
 
     private calculateDieCoordinates(
-        wafermap: WaferMap,
         mousePosition: PointCoordinates
     ): PointCoordinates | undefined {
-        if (wafermap.dataManager instanceof DataManager) {
-            const originLocation = wafermap.originLocation;
+        if (this.wafermap.dataManager instanceof DataManager) {
+            const originLocation = this.wafermap.originLocation;
             const xRoundFunction = originLocation === WaferMapOriginLocation.bottomLeft
                 || originLocation === WaferMapOriginLocation.topLeft
                 ? Math.floor
                 : Math.ceil;
             const yRoundFunction = originLocation === WaferMapOriginLocation.bottomLeft
                 || originLocation === WaferMapOriginLocation.bottomRight
-                ? Math.floor
-                : Math.ceil;
+                ? Math.ceil
+                : Math.floor;
             // go to x and y scale to get the x,y values of the die.
             const x = xRoundFunction(
-                wafermap.dataManager.invertedHorizontalScale(
-                    mousePosition.x - wafermap.dataManager.margin.left
+                this.wafermap.dataManager.horizontalScale.invert(
+                    mousePosition.x - this.wafermap.dataManager.margin.left
                 )
             );
             const y = yRoundFunction(
-                wafermap.dataManager.invertedVerticalScale(
-                    mousePosition.y - wafermap.dataManager.margin.top
+                this.wafermap.dataManager.verticalScale.invert(
+                    mousePosition.y - this.wafermap.dataManager.margin.top
                 )
             );
             return { x, y };
         }
         return undefined;
-    }
-
-    private hoversOverDie(
-        wafermap: WaferMap,
-        mousePosition: PointCoordinates
-    ): boolean {
-        const rgba = wafermap.canvasContext.getImageData(
-            mousePosition.x,
-            mousePosition.y,
-            1,
-            1
-        ).data;
-        let rgbaSum = 0;
-        for (const color of rgba) {
-            rgbaSum += color;
-        }
-        return rgbaSum > 0;
     }
 }