+++ packages/nimble-components/src/wafer-map/modules/prerendering.ts
@@ -55,7 +55,29 @@ export class Prerendering {
         const isDieRenderInfo = (
             info: DieRenderInfo | null
         ): info is DieRenderInfo => info !== null;
-        this._diesRenderInfo = this.wafermap.dies
+        if (this.wafermap.diesTable === undefined) {
+            this._diesRenderInfo = this.wafermap.dies
+                .map(die => this.computeDieRenderInfo(die))
+                .filter(isDieRenderInfo);
+            return;
+        }
+        // will chnange prerendering info for the new strategy in the following PR
+        this._diesRenderInfo = (
+            this.wafermap.diesTable.toArray() as {
+                colIndex: number,
+                rowIndex: number,
+                value: string
+            }[]
+        )
+            .map(row => {
+                // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+                return {
+                    x: row.colIndex,
+                    y: row.rowIndex,
+                    value: row.value,
+                    tags: []
+                };
+            })
             .map(die => this.computeDieRenderInfo(die))
             .filter(isDieRenderInfo);
     }
@@ -127,7 +149,7 @@ export class Prerendering {
             return '';
         }
         const label = `${value}${dieLabelsSuffix}`;
-        if (label.length > maxCharacters) {
+        if (label.length >= maxCharacters) {
             return `${label.substring(0, maxCharacters)}…`;
         }
         return label;